WebNixie/requirements.txt
flask
spidev
RPi.GPIO
